MedicalImageClassifier/requirements.txt
streamlit
onnxruntime
numpy
Pillow
torchvision
opencv-python